First revision of the shield PCB
I have a few more tweaks that I would like to make before I send this
off to be manufactured. I ended up going with a 4 layer board as it has
signal wires going all over the place since I was a bit contrained with
part placement by most of the Arduino signal pins being on the right
side of the board and wanting to place connectors on the side closest
to the part to which they connect.