Initial mriblue commit
mriblue allows GDB to debug firmware on a nRF52 device using BLE as
the communication mechanism. It uses my MRI debug monitor on the nRF52
running at IRQ level 6. This allows it to:
* Debug code that is running at IRQ level 7 or in the main thread.
* Use the BLE stack configured to run at IRQ level 5 or higher as the
  communication channel.

GDB can't communicate over BLE directly so there is also a command line
application that runs on macos to bridge between BLE and TCP/IP. This
code is found in the /mriblue/ folder.

The code in this commit is mostly just a proof of concept and doesn't
support everything need for debugging. I know it is missing:
* CTRL+C support which allows GDB to break into a running process.
* Debugging of hard faults generated at IRQ level 7 or lower.
The missing functionality will be added soon.