Start of MRI based bootloader
The mriblue_boot/ folder contains the MRI based bootloader that loads
into FLASH just above the SoftDevice, where the application would
normally reside. On startup, it initializes the BLE stack, the MRI
core, and then reads the application's ISR vector table so that it can
determine the location of its Reset_Handler. It will ask MRI to stop
at the application's Reset_Handler routine, just before jumping to it.
The developer is then free to set whatever breakpoint they want, before
main() even starts running.

I modified the firmware in the software/ folder to just blink 2 of the
LEDs on the devkit and removed all of the BLE and MRI code. This will
be the starting point of my robot firmware. This firmware image loads
into FLASH right after the bootloader.