Writes to stdout no longer waits on GDB
They just queue up a GDB 'O' (output) packet and then continue
execution of the debuggee. While sending an 'O' packet would normally
require waiting for GDB to ack it, I have modified the code to just
pretend that GDB has acked the packet and then eat the ack (or nak)
when GDB actually sends it. This means that calls from the robot code
to printf() no longer block waiting for GDB to receive the data which
can take awhile over the wireless BLE connection.