Use semihosting support to add support for mriSetDebuggerHooks()
Can't just call mriSetDebuggerHooks() like when MRI is linked into the
debuggee so use the bkpt based semihost support to pass the parameters
down to the mriblue_boot bootloader.

One catch, normally semihost calls can take awhile so the hooks are
called. But this means that they are called even when writing to
stdout (ie. printf). I have added some code on the debuggee side to
let the hooks know when a write to stdout is happening so that they
can just return and do nothing.