Add "monitor resetbreak" command support
This command allows the developer to enable or disable breaking into
the debuggee's Reset_Handler on bootup. This is enabled by default and
allows the developer to set breakpoints before the code even gets to
the program's main(). You would disable it though if you want your
code to just start and run and only halt if it crashes or you manually
attach the debugger.