Add OLED_SH1106 driver
I am now able to display things on my Pololu 1.3" 128x64 1-bit
monochrome OLED screen https://www.pololu.com/product/3760

The OLED refresh is written to take maximum advantage of the DMA
capabilities of the nRF52's SPI master peripheral. The Adafruit GFX
library just updates an in-memory frame buffer and then the refresh()
method can be called to start the copy from the in-memory frame buffer
to the one on the SH1106 itself. This copy is done a page at a time
using DMA. It only needs CPU cycles to start the next page when the
previous one completes. On the robot, the CPU can be used to update
the in-memory frame buffer when something about its state has changed
and then call refresh() to kick off this copy in the background while
the CPU does more important robot thingies.