Decrease BLE connection interval from 15ms to 7.5ms
This is the smallest connection interval allowed and normally macOS
will comply with it but I know that iOS will just drop the connection
after a minute or so when it can't agree on the interval as it wants
15ms as the shortest allowed.

macOS is sending a maximum of 8 packets during a connection interval so
decreasing it to 7.5ms results in less unused time in the connection
interval. However, I am not always seeing macOS using this shorter
connection interval for some reason.