Incorrectly treating traps as semihosting requests
The bkpt instruction range considered to be semihosting requests for
mriblue were incorrectly checked (|| instead of &&) which lead to
treating traps for things like watchpoints as a semihost request with
random results. I had a watchpoint that would hit, get treated as a
mriSetDebuggerHooks() request, corrupt the hook function pointers, and
that resulted in the device resetting.

I also fixed a memset() call to erase all of the items in the
g_whitelistPeers array.