Initial driver for TB9051FTG motor controller
Uses the PWM peripheral on the nRF52 to take care of providing the PWM
signals to both motors with minimal CPU overhead.

Also uses the SAADC peripheral to monitor the motor current and return
the current reading to the user in mA. I added a SAADCScanner object
which can be used to monitor multiple ADC channels in the background
by collecting statistics like min, max, mean, etc since the last read
call. The TB9051FTG motor driver uses it to monitor the motor currents
and in the future it can also be used to monitor other analog values
such as photocell light readings.

Also adding documents that describe the Toshiba TB9052FTG H-Bridge
chip supported by this new driver and the Pololu 20D gearmotor that I
have connected to this motor controller on my RugRover.