Patch nrfSDK Makefile.common
This patch makes the binary_name.out dependent on any libraries listed
in LIB_FILES. This means that newer versions of the MRI libraries will
cause mriblue_boot to be relinked as expected. Before I had to perform
a clean build when MRI had been updated and this was error prone.