Update PID limits with latest motor driver limits
The TB9051FTG Motor Driver doesn't allow immediate switching from one
direction to another. This update asks the motor driver for its current
limits and passes them along to the PID loop so that it knows if the
motor power will be clamped and take it into account in its integral
windup correction code.