Make sure that SAADC reading has atleast 1 sample
When sampling the SAADC channel in
DualTB9051FTGDrivers::MotorDriver::getCurrentReading(), make sure that
the reading returned has at least one sample. Based on the timing
currently used this should be guaranteed but added an assert to make
sure incase I change the SAADIC and/or PID update rates in the future.