Add Navigate module README
The main README now just redirects to the new one in the
software/Navigate folder. This Navigate specific README contains more
information about the parameters that can be tweaked for this module and
how to run the calibration processes in an attempt to calculate the
best estimates possible for them.