Increase FLASH size for mriblue_boot
It turns out that the .data initialization section was overflowing the
FLASH without producing any link errors. This commit increases the
amount of FLASH dedicated to mriblue_boot by 1 page, 4k. A future commit
will update the patch for the nrfSDK to update the common linker script
so that future .data overflows will generate a linker error.