Make TWIM/I2C operations async
Previously the I2C reads were performed in a blocking fashion which
meant that it tied up the GPIOTE high priority interrupt waiting for
them to complete. This commit refactors the code to allow the GPIOTE
interrupt handler for the falling edge of the accelerometer/magnetometer
interrupt pin to queue up the reads for the next sample of the
accelerometer, magnetometer, and gyro:
* It introduces the I2CAsync class which queues up I2C reads/writes into
a CircularBuffer object.
  * It takes a ITWIMNotification interface pointer (in pNotify) in its
  read/writeRegisters() methods. The notify() method on this interface
  will be called when the call read/write completes successfully or
  encounters an error.
  * If the pNotify parameter is NULL then the read/write call will block
  and wait for the queued up operation to complete before returning to
  the higher level caller. When non-NULL, it returns to the caller as
  soon as the operation is queued up.
  * Currently has a queue size of 8 (I2C_ASYNC_BUFFER_SIZE in
  I2CAsync.h). This might need to increase once I add code to interact
  with more I2C based sensors.
* I added a CircularBuffer class:
  * It is templated to allow for storage of custom object types within
  the queue and the size of the queue (which must be a power of 2).
  * It uses __atomic_compare_exchange_n() calls to implement a lock free
   queue that supports multiple producers and multiple consumers from
   various interrupt levels that can preempt one another.
* The FXAS21002C and FXOS8700CQ classes both allow for overlapping
getVector() calls to be made on themselves. My code tries its best to
account for this but only have single instances of members like
m_pAccelVector, m_pMagVector, m_pTemperature, etc.  This means that the
most recent I2C I/O to complete will be responsible for what the caller
finds in the buffers pointed to by these caller provided pointers. Most
of the times such multiple requests for the same sensor's measurements
 won't be found in the I2C queue at the same time. The code is just
 written to handle such overlaps gracefully if they happen to occur.
* The AdafruitPrecision9DoF class now uses the I2CAsync class instead of
interacting directly with the nrf_drv_twi_t interface. This should make
it easier to share the TWIM peripheral with other drivers in the future.