Fix potential race condition in I2CAsync
Now use a mutex to make sure that there is 1 and only 1 TWIM transfer
in progress when a new read/write operation is queued up. If there is
already a transfer in progress then the ISR will take care of issuing
the next one. If not then one producer (and only 1 producer) will be
selected to pull the oldest item from the circular buffer and submit it
to the TWIM peripheral and then the ISR can handle any more that are in
the queue.