Fix TWIM ERROR & STOPPED race condition
Sometimes the I2C master ends up sending 2 stops because the driver gets
confused when the last byte is NAKed by the slave. Sometimes this
generates an ERROR event before the STOPPED event and other times it
happens in the reverse order. I already added a fix for this in the
blocking code path of the TWIM driver in the Nordic SDK. This commit
adds a fix for the non-blocking interrupt driven version too.

I have improved the previously made linker script fix for the .data
section to make sure that the __etext symbol is given the correct
value, taking into account that the new AT usage may change alignment.