Makefile changes to reduce code size
Started at:
    text    data     bss     dec     hex filename
    90188     632    4240   95060   17354 _build/RugRover.out
Ended at:
    text    data     bss     dec     hex filename
    81736     632    4240   86608   15250 _build/RugRover.out

I had forgotten to disable exceptions as I usually do for C++ code. I
was surprised that the unwinding code doesn't take up as much space as
I remember it taking in the past.

I also disabled some of the function cloning that GCC was doing for
instances of code called with particular constant parameters. This
caused GCC to include 2 copies of functions like testNavigateRoutine()
since its bool parameter is called with both true and false values so
it included 2 versions of the code: one with the parameter set to false
and the other true.