Navigate w/ IMU and start of new Kalman filter
I updated the Navigate module so that it can now use the compass heading
determined from the IMU based Kalman filter or just use the wheel
odometry. The test results weren't that good. It seems to be mostly
related to magnetic interference in the house. Worked better outside.

Based on the poor 3D orientation Kalman filter, I have started thinking
about a new Kalman filter which just combines the yaw rate gyro output
with the wheel odometry calculated heading to see if that can generate
some improvements. This commit includes my initial work in this area
which includes the system model and code to determine what kind of
errors it generates.

I also updated some of the existing linear algebra libraries to use the
const keyword whenever possible.