Cleanup linear algebra related files
I have moved the existing linear algebra related source files into a
newly created LinearAlgebra/ folder and also renamed the existing Vector
class to Vector3. This is in preparation for the addition of Vector2 and
Matrix2x2 required for the sensor measurements of the new heading
Kalman filter that I am working on.